ppp_poems_core/poems/api/serializers.py
from rest_framework import serializers

from ppp_poems_core.poems.models import Author


class AuthorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Author
        fields = ('id', 'name', 'surname')

